requirements.txt
django
South
django-geoposition
django-allauth
MySQL-python